Remove functions from __init__.py that don't belong there
Anitya suffers from very circular imports. These are caused by defining
inter-dependent functions in various ``__init__.py``. These modules are
for package initialization. This moves all the code from
``anitya/__init__.py`` and ``anitya/lib/__init__.py`` into
``anitya/lib/utilities.py`` (not a great name, but most of the functions
inside it need to be factored out so I don't expect it to remain for too
long).

This should make working on the backend much less painful since it can
now import ``anitya.config`` without causing the entire backend to be
imported before the configuration can be imported.
